mychess/mychess/requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
chess==1.10.0
constantly==15.1.0
coverage==7.4.1
cryptography==42.0.5
daphne==4.0.0
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-htmx==1.17.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
Faker==23.2.1
flake8==6.1.0
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
mccabe==0.7.0
msgpack==1.0.5
oauthlib==3.2.2
packaging==23.2
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.7.0
pyOpenSSL==24.1.0
python-chess==1.999
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.6.3
urllib3==2.0.3
websockets==11.0.3
whitenoise==6.6.0
zope.interface==6.2
